#1 Switch between registration and login forms
