+++ app/src/main/java/com/example/gamedozor/presentation/ui/activity/MainActivity.kt
@@ -2,11 +2,17 @@ package com.example.gamedozor.presentation.ui.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
 import com.example.gamedozor.R
 
 class MainActivity : AppCompatActivity() {
+
+    private lateinit var navController: NavController
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.navMainHostActivity) as NavHostFragment
+        navController = navHostFragment.navController
     }
 }
