+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     tools:context=".presentation.ui.activity.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/fragmentContainerView"
+        android:id="@+id/navMainHostActivity"
         android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
